modify(android): Convert LanguageListActivity to utility
* Remove LanguageListActivity and KeyboardListActivity
* Keep utility method `getKeyboardsInfo()`
