+++ android/KMEA/app/src/main/AndroidManifest.xml
@@ -24,18 +24,6 @@
             android:theme="@style/Theme.AppCompat.Light.Dialog" >
         </activity>
         <activity
-            android:name="com.tavultesoft.kmea.LanguageListActivity"
-            android:configChanges="keyboard|keyboardHidden|orientation|screenLayout|screenSize|smallestScreenSize"
-            android:label="@string/app_name"
-            android:theme="@style/Theme.AppCompat.Light.Dialog" >
-        </activity>
-        <activity
-            android:name="com.tavultesoft.kmea.KeyboardListActivity"
-            android:configChanges="keyboard|keyboardHidden|orientation|screenLayout|screenSize|smallestScreenSize"
-            android:label="@string/app_name"
-            android:theme="@style/Theme.AppCompat.Light.Dialog" >
-        </activity>
-        <activity
             android:name="com.tavultesoft.kmea.KeyboardInfoActivity"
             android:configChanges="keyboard|keyboardHidden|orientation|screenLayout|screenSize|smallestScreenSize"
             android:label="@string/app_name"
